gadget/install: move kernel snap info definition to separate file
The package build on some distros (Debian Sid) removes files.
